+++ foctet-transport/src/websock.rs
@@ -49,7 +49,7 @@ pub async fn open_secure_channel_with_handshake(
     open_secure_channel_with_handshake_and_auth_config(
         session_handle,
         thresholds,
-        SessionAuthConfig::default(),
+        SessionAuthConfig::unauthenticated_for_testing(),
         TransportConfig::default(),
     )
     .await
@@ -67,7 +67,7 @@ pub async fn open_secure_channel_with_handshake_and_config(
     open_secure_channel_with_handshake_and_auth_config(
         session_handle,
         thresholds,
-        SessionAuthConfig::default(),
+        SessionAuthConfig::unauthenticated_for_testing(),
         config,
     )
     .await
@@ -135,7 +135,7 @@ pub async fn accept_secure_channel_with_handshake(
     accept_secure_channel_with_handshake_and_auth_config(
         session_handle,
         thresholds,
-        SessionAuthConfig::default(),
+        SessionAuthConfig::unauthenticated_for_testing(),
         TransportConfig::default(),
     )
     .await
@@ -153,7 +153,7 @@ pub async fn accept_secure_channel_with_handshake_and_config(
     accept_secure_channel_with_handshake_and_auth_config(
         session_handle,
         thresholds,
-        SessionAuthConfig::default(),
+        SessionAuthConfig::unauthenticated_for_testing(),
         config,
     )
     .await